timing_search.cpp now is able to produce data (execution time in function of array size) for linear seach. Gnuplot script done. Starting to generalized timing_search to produce data to a given algorithm (via command line)

+++ includes/jump_search.h
@@ -5,7 +5,7 @@
 #include <cmath> // sqrt()
 
 //=== Alias
-using value_type = int ; //!< Simple alias to help code maintenance.
+using value_type = long int ; //!< Simple alias to help code maintenance.
 
 // iterative linear search
 const value_type * lsearch( const value_type *, const value_type *, value_type );
