+++ includes/searching.h
@@ -2,7 +2,7 @@
 #define SEARCHING_H
 
 //=== Alias
-using value_type = int ; //!< Simple alias to help code maintenance.
+using value_type = long int ; //!< Simple alias to help code maintenance.
 
 // iterative linear search
 const value_type * lsearch( const value_type *, const value_type *, value_type);
